Reduce unnecesarry casting to usize

+++ chain/client-primitives/src/types.rs
@@ -485,7 +485,7 @@ impl From<near_chain_primitives::Error> for GetGasPriceError {
 pub struct NetworkInfoResponse {
     pub connected_peers: Vec<PeerInfo>,
     pub num_connected_peers: usize,
-    pub peer_max_count: u32,
+    pub peer_max_count: usize,
     pub sent_bytes_per_sec: u64,
     pub received_bytes_per_sec: u64,
     /// Accounts of known block and chunk producers from routing table.
