+++ integration-tests/src/tests/network/infinite_loop.rs
@@ -29,7 +29,7 @@ pub fn make_peer_manager(
     seed: &str,
     port: u16,
     boot_nodes: Vec<(&str, u16)>,
-    peer_max_count: u32,
+    peer_max_count: usize,
 ) -> (PeerManagerActor, PeerId, Arc<AtomicUsize>) {
     let store = create_test_store();
     let mut config = NetworkConfig::from_seed(seed, port);
